requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
cycler==0.11.0
fonttools==4.33.3
h11==0.13.0
idna==3.3
kiwisolver==1.4.2
lxml==4.7.1
matplotlib==3.5.2
numpy==1.22.2
opencv-python==4.6.0.66
outcome==1.1.0
packaging==21.3
pandas==1.4.1
Pillow==9.1.1
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
wsproto==1.0.0
